+++ src/main/java/racinggame/domain/CarInfo.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public record CarInfo(String name, int position) {
 }
