+++ src/main/java/racinggame/domain/MoveCondition.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public interface MoveCondition {
