+++ src/main/java/racinggame/domain/Round.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public class Round {
