+++ src/test/java/racinggame/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -7,6 +7,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.Car;
+import racinggame.domain.MoveCondition;
+import racinggame.domain.RacingGame;
+import racinggame.domain.Round;
 
 class RacingGameTest {
